+++ txtorcon/torstate.py
@@ -545,7 +545,7 @@ class TorState(object):
         circ.update([str(circ_id), 'EXTENDED'])
         return circ
 
-    def build_circuit(self, routers=None, using_guards=True):
+    def build_circuit(self, routers=None, using_guards=True, purpose=None):
         """
         Builds a circuit consisting of exactly the routers specified,
         in order.  This issues an EXTENDCIRCUIT call to Tor with all
@@ -586,6 +586,9 @@ class TorState(object):
                     cmd += router.decode('utf8')
                 else:
                     cmd += router.id_hex[1:]
+
+            if purpose is not None:
+                cmd += " purpose={}".format(purpose)
         d = self.protocol.queue_command(cmd)
         d.addCallback(self._find_circuit_after_extend)
         return d